+++ crates/ruma-client-api/src/delayed_events/update_delayed_event/unstable_v1.rs
@@ -55,7 +55,7 @@ mod tests {
     use std::borrow::Cow;
 
     use ruma_common::api::{
-        MatrixVersion, OutgoingRequest, SupportedVersions, auth_scheme::SendAccessToken,
+        MatrixVersion, OutgoingRequestExt as _, SupportedVersions, auth_scheme::SendAccessToken,
     };
     use serde_json::{Value as JsonValue, json};