+++ crates/ruma-client-api/src/filter/create_filter.rs
@@ -90,7 +90,8 @@ pub mod v3 {
 
             use ruma_common::{
                 api::{
-                    MatrixVersion, OutgoingRequest, SupportedVersions, auth_scheme::SendAccessToken,
+                    MatrixVersion, OutgoingRequestExt as _, SupportedVersions,
+                    auth_scheme::SendAccessToken,
                 },
                 owned_user_id,
             };