+++ crates/ruma-client-api/src/knock/knock_room.rs
@@ -61,9 +61,10 @@ pub mod v3 {
     }
 
     /// Data in the request's body.
-    #[cfg_attr(feature = "client", derive(serde::Serialize))]
+    #[doc(hidden)]
+    #[cfg_attr(feature = "client", derive(serde::Serialize, ruma_common::api::OutgoingBodyJson))]
     #[cfg_attr(feature = "server", derive(serde::Deserialize))]
-    struct RequestBody {
+    pub struct RequestBody {
         /// The reason for joining a room.
         #[serde(skip_serializing_if = "Option::is_none")]
         reason: Option<String>,
@@ -71,16 +72,16 @@ pub mod v3 {
 
     #[cfg(feature = "client")]
     impl ruma_common::api::OutgoingRequest for Request {
+        type Body = RequestBody;
         type EndpointError = Error;
         type IncomingResponse = Response;
 
-        fn try_into_http_request<T: Default + bytes::BufMut + AsRef<[u8]>>(
+        fn try_into_http_request_inner(
             self,
             base_url: &str,
-            access_token: ruma_common::api::auth_scheme::SendAccessToken<'_>,
             considering: std::borrow::Cow<'_, ruma_common::api::SupportedVersions>,
-        ) -> Result<http::Request<T>, ruma_common::api::error::IntoHttpError> {
-            use ruma_common::api::{Metadata, auth_scheme::AuthScheme};
+        ) -> Result<http::Request<Self::Body>, ruma_common::api::error::IntoHttpError> {
+            use ruma_common::api::Metadata;
 
             // Only send `server_name` if the `via` parameter is not supported by the server.
             // `via` was introduced in Matrix 1.12.
@@ -98,7 +99,7 @@ pub mod v3 {
             let query_string =
                 serde_html_form::to_string(RequestQuery { server_name, via: self.via })?;
 
-            let mut http_request = http::Request::builder()
+            let http_request = http::Request::builder()
                 .method(Self::METHOD)
                 .uri(Self::make_endpoint_url(
                     considering,
@@ -107,10 +108,7 @@ pub mod v3 {
                     &query_string,
                 )?)
                 .header(http::header::CONTENT_TYPE, ruma_common::http_headers::APPLICATION_JSON)
-                .body(ruma_common::serde::json_to_buf(&RequestBody { reason: self.reason })?)?;
-
-            Self::Authentication::add_authentication(&mut http_request, access_token)
-                .map_err(ruma_common::api::error::IntoHttpError::authentication)?;
+                .body(RequestBody { reason: self.reason })?;
 
             Ok(http_request)
         }
@@ -180,7 +178,8 @@ pub mod v3 {
 
         use ruma_common::{
             api::{
-                MatrixVersion, OutgoingRequest, SupportedVersions, auth_scheme::SendAccessToken,
+                MatrixVersion, OutgoingRequestExt as _, SupportedVersions,
+                auth_scheme::SendAccessToken,
             },
             owned_room_id, owned_server_name,
         };