+++ crates/ruma-client-api/src/session/sso_login.rs
@@ -69,7 +69,7 @@ pub mod v3 {
         use std::borrow::Cow;
 
         use ruma_common::api::{
-            MatrixVersion, OutgoingRequest, SupportedVersions, auth_scheme::SendAccessToken,
+            MatrixVersion, OutgoingRequestExt as _, SupportedVersions, auth_scheme::SendAccessToken,
         };
 
         use super::Request;