+++ crates/ruma-client-api/src/state/send_state_event.rs
@@ -128,18 +128,27 @@ pub mod v3 {
         }
     }
 
+    #[doc(hidden)]
+    #[derive(ruma_common::serde::_FakeDeriveSerde)]
+    #[cfg_attr(feature = "client", derive(serde::Serialize, ruma_common::api::OutgoingBodyJson))]
+    #[cfg_attr(feature = "server", derive(serde::Deserialize))]
+    #[serde(transparent)]
+    // attribute will go away when we update IncomingRequest to also use RequestBody
+    #[cfg_attr(not(feature = "client"), expect(dead_code))]
+    pub struct RequestBody(Raw<AnyStateEventContent>);
+
     #[cfg(feature = "client")]
     impl ruma_common::api::OutgoingRequest for Request {
+        type Body = RequestBody;
         type EndpointError = Error;
         type IncomingResponse = Response;
 
-        fn try_into_http_request<T: Default + bytes::BufMut + AsRef<[u8]>>(
+        fn try_into_http_request_inner(
             self,
             base_url: &str,
-            access_token: ruma_common::api::auth_scheme::SendAccessToken<'_>,
             considering: std::borrow::Cow<'_, ruma_common::api::SupportedVersions>,
-        ) -> Result<http::Request<T>, ruma_common::api::error::IntoHttpError> {
-            use ruma_common::api::{Metadata, auth_scheme::AuthScheme};
+        ) -> Result<http::Request<RequestBody>, ruma_common::api::error::IntoHttpError> {
+            use ruma_common::api::Metadata;
 
             let query_string = serde_html_form::to_string(RequestQuery {
                 timestamp: self.timestamp,
@@ -147,7 +156,7 @@ pub mod v3 {
                 sticky_duration_ms: self.sticky_duration_ms,
             })?;
 
-            let mut http_request = http::Request::builder()
+            let http_request = http::Request::builder()
                 .method(Self::METHOD)
                 .uri(Self::make_endpoint_url(
                     considering,
@@ -156,10 +165,7 @@ pub mod v3 {
                     &query_string,
                 )?)
                 .header(http::header::CONTENT_TYPE, ruma_common::http_headers::APPLICATION_JSON)
-                .body(ruma_common::serde::json_to_buf(&self.body)?)?;
-
-            Self::Authentication::add_authentication(&mut http_request, access_token)
-                .map_err(ruma_common::api::error::IntoHttpError::authentication)?;
+                .body(RequestBody(self.body))?;
 
             Ok(http_request)
         }
@@ -244,7 +250,9 @@ mod tests {
     use std::borrow::Cow;
 
     use ruma_common::{
-        api::{MatrixVersion, OutgoingRequest, SupportedVersions, auth_scheme::SendAccessToken},
+        api::{
+            MatrixVersion, OutgoingRequestExt as _, SupportedVersions, auth_scheme::SendAccessToken,
+        },
         owned_room_id,
     };
     use ruma_events::{EmptyStateKey, room::name::RoomNameEventContent};