+++ crates/ruma-common/tests/it/api/optional_headers.rs
@@ -4,8 +4,8 @@ use assert_matches2::assert_matches;
 use http::header::{CONTENT_DISPOSITION, LOCATION};
 use ruma_common::{
     api::{
-        IncomingRequest, IncomingResponse, MatrixVersion, OutgoingRequest, OutgoingResponse,
-        SupportedVersions, auth_scheme::NoAuthentication, request, response,
+        IncomingRequest, IncomingResponse, MatrixVersion, OutgoingRequestExt as _,
+        OutgoingResponse, SupportedVersions, auth_scheme::NoAuthentication, request, response,
     },
     http_headers::{ContentDisposition, ContentDispositionType},
     metadata,